refactored view full collection to just return the collection
