+++ collection.txt
@@ -1,0 +1,2 @@
+1 Teysa, Envoy of Ghosts (c21)
+1 Rancor (2x2)
